skillforge_v2/requirements.txt
Django==5.0.6
djangorestframework==3.17.1
requests==2.32.3
# Standard additions for Django production and utility
whitenoise==6.7.0
gunicorn==22.0.0
Pillow==10.4.0
sqlparse==0.5.1
